refactor: modularizar dominio orders em api/hooks/ui/types
