+++ src/app/routes/index.ts
@@ -20,7 +20,7 @@ const Dashboard = lazy(() => import("./dashboard"))
 const DashboardAddOrganization = lazy(() => import("./dashboard/addOrg"))
 const DashboardHome = lazy(() => import("./dashboard/pages/dashboard-home"))
 const Account = lazy(() => import("./dashboard/pages/account"))
-const Orders = lazy(() => import("./dashboard/pages/orders"))
+const Orders = lazy(() => import("@/domains/orders/ui/pages/orders-page"))
 const POS = lazy(() => import("./dashboard/pages/pos"))
 const Settings = lazy(() => import("./dashboard/pages/settings"))
 const Tables = lazy(() => import("./dashboard/pages/tables"))
@@ -40,12 +40,12 @@ const Inventory = lazy(() =>
 	}))
 )
 const Kitchen = lazy(() =>
-	import("./dashboard/pages/kitchen").then((module) => ({
+	import("@/domains/orders/ui/pages/kitchen-page").then((module) => ({
 		default: module.KitchenPage,
 	}))
 )
 const Delivery = lazy(() =>
-	import("./dashboard/pages/delivery").then((module) => ({
+	import("@/domains/orders/ui/pages/delivery-page").then((module) => ({
 		default: module.DeliveryPage,
 	}))
 )
