+++ src/domains/orders/model/order-store.ts
@@ -1,0 +1,50 @@
+import { create } from "zustand"
+import type { Order, OrderStatus, OrderType } from "../types/order.types"
+
+interface OrderStore {
+	selectedOrder: Order | null
+	filter: {
+		status: OrderStatus | "all" | "approval_pending"
+		type: OrderType | "all"
+		search: string
+	}
+	setFilter: (key: "status" | "type" | "search", value: string) => void
+	selectOrder: (order: Order | null) => void
+}
+
+export const useOrderStore = create<OrderStore>((set) => ({
+	selectedOrder: null,
+	filter: {
+		status: "all",
+		type: "all",
+		search: "",
+	},
+	setFilter: (key, value) =>
+		set((state) => ({
+			filter: { ...state.filter, [key]: value },
+		})),
+	selectOrder: (order) => set({ selectedOrder: order }),
+}))
+
+export const useOrderSelectors = () => {
+	const selectedOrder = useOrderStore((state) => state.selectedOrder)
+	const filter = useOrderStore((state) => state.filter)
+
+	return {
+		selectedOrder,
+		filter,
+	}
+}
+
+export const useOrderActions = () => {
+	const setFilter = useOrderStore((state) => state.setFilter)
+	const selectOrder = useOrderStore((state) => state.selectOrder)
+
+	return {
+		setFilter,
+		selectOrder,
+	}
+}
+
+export const useSelectedOrder = useOrderSelectors
+export const useOrdersActions = useOrderActions
