+++ src/domains/orders/store/order-store.ts
@@ -1,50 +1,13 @@
-import { create } from "zustand"
-import type { Order, OrderStatus, OrderType } from "@/types/dashboard"
+import {
+	useOrderActions as domainUseOrderActions,
+	useOrderSelectors as domainUseOrderSelectors,
+	useOrderStore as domainUseOrderStore,
+	useOrdersActions as domainUseOrdersActions,
+	useSelectedOrder as domainUseSelectedOrder,
+} from "../model/order-store"
 
-interface OrderStore {
-	selectedOrder: Order | null
-	filter: {
-		status: OrderStatus | "all" | "approval_pending"
-		type: OrderType | "all"
-		search: string
-	}
-	setFilter: (key: "status" | "type" | "search", value: string) => void
-	selectOrder: (order: Order | null) => void
-}
-
-export const useOrderStore = create<OrderStore>((set) => ({
-	selectedOrder: null,
-	filter: {
-		status: "all",
-		type: "all",
-		search: "",
-	},
-	setFilter: (key, value) =>
-		set((state) => ({
-			filter: { ...state.filter, [key]: value },
-		})),
-	selectOrder: (order) => set({ selectedOrder: order }),
-}))
-
-export const useOrderSelectors = () => {
-	const selectedOrder = useOrderStore((state) => state.selectedOrder)
-	const filter = useOrderStore((state) => state.filter)
-
-	return {
-		selectedOrder,
-		filter,
-	}
-}
-
-export const useOrderActions = () => {
-	const setFilter = useOrderStore((state) => state.setFilter)
-	const selectOrder = useOrderStore((state) => state.selectOrder)
-
-	return {
-		setFilter,
-		selectOrder,
-	}
-}
-
-export const useSelectedOrder = useOrderSelectors
-export const useOrdersActions = useOrderActions
+export const useOrderStore = domainUseOrderStore
+export const useOrderSelectors = domainUseOrderSelectors
+export const useOrderActions = domainUseOrderActions
+export const useSelectedOrder = domainUseSelectedOrder
+export const useOrdersActions = domainUseOrdersActions
