+++ src/hooks/use-orders.ts
@@ -1,38 +1,33 @@
-import { useQuery } from "@tanstack/react-query"
-import { getOrderStats, getOrders } from "@/services/orders-service"
-import type { OrderFilter } from "@/types/orders"
+import { ordersQueryKeys as domainOrdersQueryKeys } from "@/domains/orders/hooks/orders-query-keys"
+import {
+	useOrderStats as domainUseOrderStats,
+	useOrders as domainUseOrders,
+} from "@/domains/orders/hooks/use-orders"
+import type { OrderFilter } from "@/domains/orders/types/order-query.types"
 
 export const orderQueryKeys = {
-	all: () => ["orders"],
-	getOrders: (filters?: OrderFilter, page?: number, count?: number) => [
-		...orderQueryKeys.all(),
-		"list",
-		{ filters },
-		{ page },
-		{ count },
-	],
-	getOrderById: (orderId: string) => [
-		...orderQueryKeys.all(),
-		"one",
-		{ orderId },
-	],
-	getOrderStats: () => [...orderQueryKeys.all(), "stats"],
+	all: () => [...domainOrdersQueryKeys.all],
+	getOrders: (filters?: OrderFilter, page?: number, count?: number) =>
+		domainOrdersQueryKeys.list(null, filters, page, count),
+	getOrderById: (orderId: string) =>
+		domainOrdersQueryKeys.detail(null, orderId),
+	getOrderStats: () => domainOrdersQueryKeys.stats(null),
 }
 
 export function useOrders(
 	filters?: OrderFilter,
 	page?: number,
-	count?: number
+	count?: number,
+	organizationId?: string | null
 ) {
-	return useQuery({
-		queryKey: orderQueryKeys.getOrders(filters, page, count),
-		queryFn: ({ signal }) => getOrders(filters, page, count, signal),
+	return domainUseOrders({
+		organizationId: organizationId ?? null,
+		filters,
+		page,
+		count,
 	})
 }
 
-export function useOrderStats() {
-	return useQuery({
-		queryKey: orderQueryKeys.getOrderStats(),
-		queryFn: ({ signal }) => getOrderStats(signal),
-	})
+export function useOrderStats(organizationId?: string | null) {
+	return domainUseOrderStats(organizationId ?? null)
 }
